Add fs.write module and tests

+++ libfs/disk.c
@@ -51,8 +51,8 @@ int block_disk_open(const char *diskname)
 
 	/* The disk image's size should be a multiple of the block size */
 	if (st.st_size % BLOCK_SIZE != 0) {
-		block_error("size '%zu' is not multiple of '%d'",
-			    st.st_size, BLOCK_SIZE);
+		block_error("size '%lld' is not multiple of '%d'",
+			    (long long)st.st_size, BLOCK_SIZE);
 		return -1;
 	}
